tarea  terminada

+++ Unidad-4/tareas/Interfaces/src/main/java/ies/puerto/interfaz/IFormaGeometrica.java
@@ -1,5 +1,5 @@
 package ies.puerto.interfaz;
 
 public interface IFormaGeometrica {
-    double calcularArea();
+    float calcularArea();
 }
